pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "img2webp"
version = "0.1.0"
description = "Convert PNG/JPG images to optimized WebP via CLI"
requires-python = ">=3.10,<3.13"
dependencies = [
    "typer[all]>=0.24.2",
    "Pillow>=12.2.0",
    "rich>=15.0.0",
    "pillow-heif>=1.3.0",
    "PyMuPDF>=1.27.2",
    "scipy>=1.11",
    "numpy>=2.0.0",
]

[project.scripts]
img2webp = "img2webp.cli:app"
heic2jpg = "heic2jpg.cli:app"
pdf2jpeg = "pdf2jpeg.cli:app"
pdfcompress = "pdfcompress.cli:app"
vid2audio = "vid2audio.cli:app"
rmbg = "rm_bg.cli:app"
[tool.setuptools.packages.find]
where = ["."]
include = ["img2webp*", "heic2jpg*", "pdf2jpeg*", "pdfcompress*", "rm_bg*"]
